requirements.txt
bs4==0.0.1
requests>=2.20.0
pycryptodome==3.6.6
python-dateutil
